glslang: Rename and move cmake export
CMake when searching for a package in config mode will by default look for
a file called <name>Config.cmake or <lower-case-name>-config.cmake.
The previous naming would force the user to specify the name to
glslangTargets. The name is therefore changed to glslangConfig.cmake.

On Unix, cmake will look for the config files in the following directories:
<prefix>/(lib/<arch>|lib|share)/cmake/<name>*/          (U)
<prefix>/(lib/<arch>|lib|share)/<name>*/                (U)
<prefix>/(lib/<arch>|lib|share)/<name>*/(cmake|CMake)/  (U)
Previously the destination directory was <prefix>/lib/cmake which would not
be found.
The destination directory now includes the package name.

A namespace is added to be able to specify components.